+++ src/main/java/com/telepathicgrunt/the_bumblezone/mixin/entities/FloatGoalMixin.java
@@ -1,0 +1,54 @@
+package com.telepathicgrunt.the_bumblezone.mixin.entities;
+
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.telepathicgrunt.the_bumblezone.blocks.EmptyHoneycombBrood;
+import com.telepathicgrunt.the_bumblezone.blocks.FilledPorousHoneycomb;
+import com.telepathicgrunt.the_bumblezone.blocks.PorousHoneycomb;
+import com.telepathicgrunt.the_bumblezone.entities.EntityTeleportationBackend;
+import com.telepathicgrunt.the_bumblezone.items.StinglessBeeHelmet;
+import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
+import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
+import net.minecraft.core.BlockPos;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.ai.goal.FloatGoal;
+import net.minecraft.world.entity.animal.Bee;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Fluid;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.phys.AABB;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.Set;
+
+@Mixin(FloatGoal.class)
+public abstract class FloatGoalMixin {
+
+    @Final
+    @Shadow
+    private Mob mob;
+
+    // let Bumblezone fluids have entities float in its honey
+    @ModifyExpressionValue(method = "canUse()Z",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Mob;getFluidHeight(Lnet/minecraft/tags/TagKey;)D"))
+    private double thebumblezone_fluidFloating(double originalFluidHeight) {
+        double newFluidHeight = mob.getFluidHeight(BzTags.SPECIAL_HONEY_LIKE);
+        if (newFluidHeight > originalFluidHeight) {
+            return newFluidHeight;
+        }
+        return originalFluidHeight;
+    }
+}