+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Wrath of the Hive will not apply to players in Peaceful mode now.
 ##### Fluids:
 Fixed it so Bees that spend too long in Honey Fluid and Royal Jelly Fluid will drown properly.
 
+Fixed Honey Fluid and Royal Jelly Fluid so mobs can float on top of it now.
+
 ##### Lang:
 Added a few extra details on some blocks in the JEI/REI/EMI item descriptions.